Feature: Bert e2e pipeline
