+++ bert/bert_train.py
@@ -19,7 +19,7 @@ from sklearn.metrics import accuracy_score
 from sklearn.model_selection import train_test_split
 from torch import nn
 from torch.utils.data import DataLoader, Dataset
-from torchtext.datasets.text_classification import URLS
+import torchtext.datasets as td
 from torchtext.utils import download_from_url, extract_archive
 from transformers import AdamW, BertModel, BertTokenizer
 from pytorch_lightning.loggers import TensorBoardLogger
@@ -28,179 +28,6 @@ from pathlib import Path
 import boto3
 from botocore.exceptions import ClientError
 import shutil
-
-
-class NewsDataset(Dataset):
-    def __init__(self, reviews, targets, tokenizer, max_length):
-        """
-        Performs initialization of tokenizer
-
-        :param reviews: AG news text
-        :param targets: labels
-        :param tokenizer: bert tokenizer
-        :param max_length: maximum length of the news text
-
-        """
-        self.reviews = reviews
-        self.targets = targets
-        self.tokenizer = tokenizer
-        self.max_length = max_length
-
-    def __len__(self):
-        """
-        :return: returns the number of datapoints in the dataframe
-
-        """
-        return len(self.reviews)
-
-    def __getitem__(self, item):
-        """
-        Returns the review text and the targets of the specified item
-
-        :param item: Index of sample review
-
-        :return: Returns the dictionary of review text, input ids, attention mask, targets
-        """
-        review = str(self.reviews[item])
-        target = self.targets[item]
-
-        encoding = self.tokenizer.encode_plus(
-            review,
-            add_special_tokens=True,
-            max_length=self.max_length,
-            return_token_type_ids=False,
-            padding="max_length",
-            return_attention_mask=True,
-            return_tensors="pt",
-            truncation=True,
-        )
-
-        return {
-            "review_text": review,
-            "input_ids": encoding["input_ids"].flatten(),
-            "attention_mask": encoding["attention_mask"].flatten(),
-            "targets": torch.tensor(target, dtype=torch.long),
-        }
-
-
-class BertDataModule(pl.LightningDataModule):
-    def __init__(self, **kwargs):
-        """
-        Initialization of inherited lightning data module
-        """
-        super(BertDataModule, self).__init__()
-        self.PRE_TRAINED_MODEL_NAME = "bert-base-uncased"
-        self.df_train = None
-        self.df_val = None
-        self.df_test = None
-        self.train_data_loader = None
-        self.val_data_loader = None
-        self.test_data_loader = None
-        self.MAX_LEN = 100
-        self.encoding = None
-        self.tokenizer = None
-        self.args = kwargs
-
-    def prepare_data(self):
-        """
-        Implementation of abstract class
-        """
-
-    @staticmethod
-    def process_label(rating):
-        rating = int(rating)
-        return rating - 1
-
-    def setup(self, stage=None):
-        """
-        Downloads the data, parse it and split the data into train, test, validation data
-
-        :param stage: Stage - training or testing
-        """
-
-        num_samples = self.args["num_samples"]
-
-        data_path = self.args["train_glob"]
-
-        print("\n\nTRAIN GLOB")
-        print(data_path)
-        print("\n\n")
-
-        df_parquet = pq.ParquetDataset(self.args["train_glob"])
-
-        df = df_parquet.read_pandas().to_pandas()
-
-        df.columns = ["label", "title", "description"]
-        df.sample(frac=1)
-        df = df.iloc[:num_samples]
-
-        df["label"] = df.label.apply(self.process_label)
-
-        self.tokenizer = BertTokenizer.from_pretrained(self.PRE_TRAINED_MODEL_NAME)
-
-        RANDOM_SEED = 42
-        np.random.seed(RANDOM_SEED)
-        torch.manual_seed(RANDOM_SEED)
-
-        self.df_train, self.df_test = train_test_split(
-            df, test_size=0.1, random_state=RANDOM_SEED, stratify=df["label"]
-        )
-        self.df_val, self.df_test = train_test_split(
-            self.df_test,
-            test_size=0.5,
-            random_state=RANDOM_SEED,
-            stratify=self.df_test["label"],
-        )
-
-    def create_data_loader(self, df, tokenizer, max_len, batch_size):
-        """
-        Generic data loader function
-
-        :param df: Input dataframe
-        :param tokenizer: bert tokenizer
-        :param max_len: Max length of the news datapoint
-        :param batch_size: Batch size for training
-
-        :return: Returns the constructed dataloader
-        """
-        ds = NewsDataset(
-            reviews=df.description.to_numpy(),
-            targets=df.label.to_numpy(),
-            tokenizer=tokenizer,
-            max_length=max_len,
-        )
-
-        return DataLoader(
-            ds, batch_size=self.args["batch_size"], num_workers=self.args["num_workers"]
-        )
-
-    def train_dataloader(self):
-        """
-        :return: output - Train data loader for the given input
-        """
-        self.train_data_loader = self.create_data_loader(
-            self.df_train, self.tokenizer, self.MAX_LEN, self.args["batch_size"]
-        )
-        return self.train_data_loader
-
-    def val_dataloader(self):
-        """
-        :return: output - Validation data loader for the given input
-        """
-        self.val_data_loader = self.create_data_loader(
-            self.df_val, self.tokenizer, self.MAX_LEN, self.args["batch_size"]
-        )
-        return self.val_data_loader
-
-    def test_dataloader(self):
-        """
-        :return: output - Test data loader for the given input
-        """
-        self.test_data_loader = self.create_data_loader(
-            self.df_test, self.tokenizer, self.MAX_LEN, self.args["batch_size"]
-        )
-        return self.test_data_loader
-
 
 class BertNewsClassifier(pl.LightningModule):
     def __init__(self, **kwargs):
@@ -218,6 +45,7 @@ class BertNewsClassifier(pl.LightningModule):
         n_classes = len(self.class_names)
 
         self.fc1 = nn.Linear(self.bert_model.config.hidden_size, 512)
+        self.embedding = nn.EmbeddingBag(1308844, 32, sparse=True)
         self.out = nn.Linear(512, n_classes)
 
         self.scheduler = None
@@ -254,8 +82,9 @@ class BertNewsClassifier(pl.LightningModule):
         attention_mask = train_batch["attention_mask"].to(self.device)
         targets = train_batch["targets"].to(self.device)
         output = self.forward(input_ids, attention_mask)
+        _, y_hat = torch.max(output, dim=1)
         loss = F.cross_entropy(output, targets)
-        self.train_acc(output, targets)
+        self.train_acc(y_hat, targets)
         self.log("train_acc", self.train_acc.compute())
         self.log("train_loss", loss)
         return {"loss": loss, "acc": self.train_acc.compute()}
@@ -293,8 +122,9 @@ class BertNewsClassifier(pl.LightningModule):
         attention_mask = val_batch["attention_mask"].to(self.device)
         targets = val_batch["targets"].to(self.device)
         output = self.forward(input_ids, attention_mask)
+        _, y_hat = torch.max(output, dim=1)
         loss = F.cross_entropy(output, targets)
-        self.val_acc(output, targets)
+        self.val_acc(y_hat, targets)
         self.log("val_acc", self.val_acc.compute())
         self.log("val_loss", loss, sync_dist=True)
         return {"val_step_loss": loss, "acc": self.val_acc.compute()}
@@ -361,6 +191,7 @@ def train_model(
         "accelerator": accelerator,
     }
 
+    from bert_datamodule import BertDataModule
     dm = BertDataModule(**dict_args)
     dm.prepare_data()
     dm.setup(stage="fit")
@@ -384,20 +215,21 @@ def train_model(
         save_top_k=1,
         verbose=True,
         monitor="val_loss",
-        mode="min",
-        prefix="",
+        mode="min"
     )
     lr_logger = LearningRateMonitor()
 
     trainer = pl.Trainer(
         logger=tboard,
         accelerator=accelerator,
-        callbacks=[lr_logger, early_stopping],
-        checkpoint_callback=checkpoint_callback,
+        callbacks=[lr_logger, early_stopping, checkpoint_callback],
+        checkpoint_callback=True,
         max_epochs=max_epochs,
     )
+
     trainer.fit(model, dm)
     trainer.test()
+    torch.save(model.state_dict(), os.path.join(model_save_path, 'bert.pth'))
 
 
 if __name__ == "__main__":
